Securely Download Google Sheet Data to Workbook
If you have a form set up to send data to Google sheets and you need to manipulate it with the full power of Excel, then you can use the Google SQL API to automatically pull in the data from the worksheet for you. Using this module you can avoid having to manually download the data or manually copy and paste it to the workbook. You don't have to publish the Google sheet to the web, but can instead use Google SQL system.