Create a Collection of Selected Terms in a Listbox
If you have included a list box in a Userform or a sheet, where the user selects from the options you give them, then this function is just the thing for you. All this function needs is the Listbox object you would like it to search. If the user does not make a selection from the Listbox, the function will return an error message and end all macro processes.