Create a Collection from the Items In a Sheet
Ever need to create a collection from the items in a sheet? Ever need only the items in a column that are not blank ("" or vbNullString)? Ever need to collect only the unique items from a column, such as from a product list in a database? I know I have. This module takes care of all of these problems in a flash. I hope you get some good use out of it.