Convert a Collection to An Array
So, you have your collection, but realize that for whatever reason you now need an array. Rather than go back to square one and create the array through the same (possibly extensive) processes as those that created the collection, you can just convert the collection to an array with this module.