Convert Number to Letter
You may often find the need to loop through the columns of a workbook, but you have the problem that you need to convert the numbers to letters (depending on how you are referring to the cells); this function solves this problem. Hopefully, you can find a few other uses for it that haven't been thought up yet.