Safeguard Against Starting a Long Process
If you have ever built a Userform or an interface on a sheet in Excel, you know that more than a few times you and your users have let their fingers slipped and started a long process that may damage all the progress that has already been made. Place this function as the end of all lengthy processes, type what process the user was about to start, and a MsgBox asking the user if they want to proceed will popup.